+++ execution-engine/engine-core/src/engine_state/genesis.rs
@@ -6,7 +6,7 @@ use rand::{
     Rng,
 };
 
-use engine_shared::{motes::Motes, newtypes::Blake2bHash, transform::TypeMismatch};
+use engine_shared::{motes::Motes, newtypes::Blake2bHash, TypeMismatch};
 use engine_storage::global_state::CommitResult;
 use engine_wasm_prep::wasm_costs::WasmCosts;
 use types::{account::PublicKey, bytesrepr, Key, ProtocolVersion, U512};
