+++ execution-engine/engine-core/src/engine_state/upgrade.rs
@@ -1,6 +1,6 @@
 use std::fmt;
 
-use engine_shared::{newtypes::Blake2bHash, transform::TypeMismatch};
+use engine_shared::{newtypes::Blake2bHash, TypeMismatch};
 use engine_storage::global_state::CommitResult;
 use engine_wasm_prep::wasm_costs::WasmCosts;
 use types::{bytesrepr, Key, ProtocolVersion};
