+++ execution-engine/engine-core/src/tracking_copy/ext.rs
@@ -2,7 +2,7 @@ use std::convert::TryInto;
 
 use engine_shared::{
     account::Account, contract::Contract, motes::Motes, newtypes::CorrelationId,
-    stored_value::StoredValue, transform::TypeMismatch,
+    stored_value::StoredValue, TypeMismatch,
 };
 use engine_storage::global_state::StateReader;
 use types::{bytesrepr::ToBytes, CLValue, Key, URef, U512};
