+++ execution-engine/engine-grpc-server/src/engine_server/mappings/transforms/error.rs
@@ -1,6 +1,6 @@
 use std::convert::TryFrom;
 
-use engine_shared::transform::{self, TypeMismatch};
+use engine_shared::{transform, TypeMismatch};
 
 use crate::engine_server::{
     mappings::ParsingError,
