+++ execution-engine/engine-shared/src/lib.rs
@@ -13,4 +13,7 @@ pub mod socket;
 pub mod stored_value;
 pub mod test_utils;
 pub mod transform;
+mod type_mismatch;
 pub mod utils;
+
+pub use type_mismatch::TypeMismatch;
