+++ execution-engine/engine-shared/src/transform.rs
@@ -14,19 +14,7 @@ use types::{
     CLType, CLTyped, CLValue, CLValueError, Key, U128, U256, U512,
 };
 
-use crate::stored_value::StoredValue;
-
-#[derive(PartialEq, Eq, Debug, Clone)]
-pub struct TypeMismatch {
-    pub expected: String,
-    pub found: String,
-}
-
-impl TypeMismatch {
-    pub fn new(expected: String, found: String) -> TypeMismatch {
-        TypeMismatch { expected, found }
-    }
-}
+use crate::{stored_value::StoredValue, TypeMismatch};
 
 /// Error type for applying and combining transforms. A `TypeMismatch`
 /// occurs when a transform cannot be applied because the types are
